kie-issues#1801: Fix ClassNotFoundException when Registering classes for Reflection
