+++ api/kogito-jobs-service-api/src/main/java/org/kie/kogito/jobs/service/api/event/JobCloudEvent.java
@@ -23,8 +23,8 @@ import java.time.OffsetDateTime;
 import java.util.Objects;
 import java.util.Set;
 
-import org.kie.kogito.jobs.service.api.serlialization.SpecVersionDeserializer;
-import org.kie.kogito.jobs.service.api.serlialization.SpecVersionSerializer;
+import org.kie.kogito.jobs.service.api.serialization.SpecVersionDeserializer;
+import org.kie.kogito.jobs.service.api.serialization.SpecVersionSerializer;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
