+++ api/kogito-jobs-service-api/src/main/java/org/kie/kogito/jobs/service/api/event/serialization/JobCloudEventSerializer.java
@@ -23,8 +23,8 @@ import org.kie.kogito.jobs.service.api.event.JobCloudEvent;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import static org.kie.kogito.jobs.service.api.serlialization.SerializationUtils.DEFAULT_OBJECT_MAPPER;
-import static org.kie.kogito.jobs.service.api.serlialization.SerializationUtils.registerDescriptors;
+import static org.kie.kogito.jobs.service.api.serialization.SerializationUtils.DEFAULT_OBJECT_MAPPER;
+import static org.kie.kogito.jobs.service.api.serialization.SerializationUtils.registerDescriptors;
 
 public class JobCloudEventSerializer {
