+++ api/kogito-jobs-service-api/src/main/java/org/kie/kogito/jobs/service/api/serialization/DeserializationException.java
@@ -16,20 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.jobs.service.api.serlialization;
+package org.kie.kogito.jobs.service.api.serialization;
 
-import java.io.IOException;
+public class DeserializationException extends RuntimeException {
 
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
+    public DeserializationException(String message) {
+        super(message);
+    }
 
-import io.cloudevents.SpecVersion;
-
-public class SpecVersionSerializer extends JsonSerializer<SpecVersion> {
-
-    @Override
-    public void serialize(SpecVersion specVersion, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
-        jsonGenerator.writeString(specVersion.toString());
+    public DeserializationException(String message, Throwable cause) {
+        super(message, cause);
     }
 }
