+++ api/kogito-jobs-service-api/src/main/java/org/kie/kogito/jobs/service/api/serialization/SpecVersionDeserializer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.jobs.service.api.serlialization;
+package org.kie.kogito.jobs.service.api.serialization;
 
 import java.io.IOException;
