+++ api/kogito-jobs-service-api/src/main/java/org/kie/kogito/jobs/service/api/utils/ReflectionUtils.java
@@ -39,8 +39,8 @@ import org.kie.kogito.jobs.service.api.recipient.sink.serialization.ContentModeD
 import org.kie.kogito.jobs.service.api.recipient.sink.serialization.ContentModeSerializer;
 import org.kie.kogito.jobs.service.api.schedule.cron.CronSchedule;
 import org.kie.kogito.jobs.service.api.schedule.timer.TimerSchedule;
-import org.kie.kogito.jobs.service.api.serlialization.SpecVersionDeserializer;
-import org.kie.kogito.jobs.service.api.serlialization.SpecVersionSerializer;
+import org.kie.kogito.jobs.service.api.serialization.SpecVersionDeserializer;
+import org.kie.kogito.jobs.service.api.serialization.SpecVersionSerializer;
 
 public class ReflectionUtils {
